% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_downsampling_DGEanalysis.r
\name{bulk_downsampling_DGEanalysis}
\alias{bulk_downsampling_DGEanalysis}
\title{Downsample the dataset, based either on the individuals or cells, and run DE analysis on each downsampled output. Save results in a dataframe}
\usage{
bulk_downsampling_DGEanalysis(
  SCEs,
  dataset_names,
  celltype_correspondence,
  sampled = "individuals",
  sampleIDs = "donor_id",
  assay_names = "counts",
  celltypeIDs = "cell_type",
  output_path = getwd(),
  pvalue = 0.05,
  Nperms = 20
)
}
\arguments{
\item{SCEs}{A list of SingleCellExperiment (SCE) objects, each representing a scRNA-seq dataset.}

\item{dataset_names}{A vector of names corresponding to each dataset (as you would like them to appear in output plots).}

\item{celltype_correspondence}{A named vector that maps a standard cell type label (e.g., \code{"Endo"}, \code{"Micro"}) to how that cell type appears in each dataset. Use \code{NA} if the cell type is not present in a given dataset.}

\item{sampled}{Specifies the unit of down-sampling. Can be either \code{"individuals"} or \code{"cells"}, depending on whether the analysis downsamples across samples or cells.}

\item{sampleIDs}{A character vector specifying the column name in each SCE that represents sample or donor IDs (in order of SCEs).}

\item{assay_names}{A character vector specifying the assay names in each SCE that will be used for the analysis (in order of SCEs). Default is a vector with all entries \code{"counts"}, which uses the count assay in each SCE.}

\item{celltypeIDs}{A character vector specifying the column name in each SCE that denotes cell type identity (in order of SCEs).}

\item{output_path}{A clean directory path where down-sampled outputs and plots will be saved (should contain no subdirectories).}

\item{pvalue}{P-value threshold for defining DEGs in the bulk dataset.}

\item{Nperms}{Number of permutations to perform for each down-sampling level. Default is 20.
Saves DGE analysis output in the correct directory, to be used by other bulk analysis functions}
}
\description{
Downsample the dataset, based either on the individuals or cells, and run DE analysis on each downsampled output. Save results in a dataframe
}
