% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_downsampling_range.r
\name{bulk_downsampling_range}
\alias{bulk_downsampling_range}
\title{Obtain the range of values to downsample at (either for individuals, or mean number of cells per individual), for bulk analysis}
\usage{
bulk_downsampling_range(
  SCEs,
  celltype_correspondence,
  sampled = "individuals",
  sampleIDs = "donor_id",
  celltypeIDs = "cell_type"
)
}
\arguments{
\item{SCEs}{A list of SingleCellExperiment (SCE) objects, each representing a scRNA-seq dataset.}

\item{celltype_correspondence}{A named vector that maps a standard cell type label (e.g., \code{"Endo"}, \code{"Micro"}) to how that cell type appears in each dataset. Use \code{NA} if the cell type is not present in a given dataset.}

\item{sampled}{Specifies the unit of down-sampling. Can be either \code{"individuals"} or \code{"cells"}, depending on whether the analysis downsamples across samples or cells.}

\item{sampleIDs}{A character vector specifying the column name in each SCE that represents sample or donor IDs (in order of SCEs).}

\item{celltypeIDs}{A character vector specifying the column name in each SCE that denotes cell type identity (in order of SCEs).}
}
\value{
list containing values which the data will be downsampled at, in ascending order
}
\description{
Obtain the range of values to downsample at (either for individuals, or mean number of cells per individual), for bulk analysis
}
