% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_downsampled_corr.r
\name{compute_downsampled_corr}
\alias{compute_downsampled_corr}
\title{For a given down-sampled DGE analysis output, computes the correlation of the log-foldchange of the DEGs (at specified p-value) for a given dataset (celltype)}
\usage{
compute_downsampled_corr(allstudies, sampled = "individuals")
}
\arguments{
\item{allstudies}{a list containing DGE analysis outputs (subset to "celltype_all_genes" and the specified cell type) at each down-sampled point}

\item{sampled}{Specifies the unit of down-sampling. Can be either \code{"individuals"} or \code{"cells"}, depending on whether the analysis downsamples across samples or cells.}
}
\value{
correlation matrix, plot and the number of DEGs at the specified p-value
}
\description{
For a given down-sampled DGE analysis output, computes the correlation of the log-foldchange of the DEGs (at specified p-value) for a given dataset (celltype)
}
