% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.r
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Runs correlation analysis pipeline}
\usage{
correlation_analysis(
  dataset_name = "placeholder",
  allstudies = "placeholder",
  celltypes = "placeholder",
  pvals = c(0.05, 0.025, 0.01, 0.001, 1e-04),
  data_names = "placeholder",
  alphaval = "placeholder",
  numPerms = "placeholder",
  numSubsets = "placeholder",
  output_path = getwd()
)
}
\arguments{
\item{dataset_name}{name of the dataset used to select significant DEGs from (specified as a string, name as in allstudies)}

\item{allstudies}{a list containing all the datasets (as SCE objects)}

\item{celltypes}{a list containing the celltypes to compute mean correlation across}

\item{pvals}{list of p-value cut-offs which will be used to select DEGs}

\item{data_names}{names of the datasets as they appear in the correlation plot}

\item{alphaval}{(alpha) transparency of the non-mean boxplots}

\item{numPerms}{number of random permutations of the dataset used to select significant DEGs from}

\item{numSubsets}{number of pairs of random subsets of the dataset used to select significant DEGs from}

\item{output_path}{base path in which outputs will be stored
Saves all plots and DGE analysis outputs in the appropriate directories}
}
\description{
Runs correlation analysis pipeline
}
