% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling_corrplots.r
\name{downsampling_corrplots}
\alias{downsampling_corrplots}
\title{Create correlation plots of the effect sizes of the top 1000 and top 500 DEGs}
\usage{
downsampling_corrplots(
  data,
  range_downsampled = "placeholder",
  output_path = getwd(),
  inpath = "placeholder",
  sampled = "individuals",
  sampleID = "donor_id",
  celltypeID = "cell_type",
  coeff = "male",
  Nperms = 20,
  y = NULL,
  region = "single_region",
  control = NULL,
  pval_adjust_method = "BH",
  rmv_zero_count_genes = TRUE
)
}
\arguments{
\item{data}{the input data (should be an SCE object)}

\item{range_downsampled}{vector or list containing values which the data will be downsampled at, in ascending order}

\item{output_path}{base path in which outputs will be stored}

\item{inpath}{base path where downsampled DGE analysis output folders are stored (taken to be output_path if not provided)}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{sampleID}{sample ID}

\item{celltypeID}{cell type ID}

\item{coeff}{which coefficient to carry out DE analysis with respect to}

\item{Nperms}{number of subsets created when downsampling at each level}

\item{y}{the column name in the SCE object for the return variable e.g. "diagnosis" - Case or disease. Default is the last variable in the design formula. y can be discrete (logistic regression) or continuous (linear regression)}

\item{region}{the column name in the SCE object for the study region. If there are multiple regions in the study (for example two brain regions). Pseudobulk values can be derived separately. Default is "single_region" which will not split by region.}

\item{control}{character specifying which control level for the differential expression analysis e.g. in a case/control/other study use "control" in the y column to compare against. NOTE only need to specify if more than two groups in y, leave as default value for two groups or continuous y. Default is NULL.}

\item{pval_adjust_method}{the adjustment method for the p-value in the differential expression analysis. Default is benjamini hochberg "BH". See  stats::p.adjust for available options}

\item{rmv_zero_count_genes}{whether genes with no count values in any cell should be removed. Default is TRUE
Saves all plots in the appropriate directory}
}
\description{
Create correlation plots of the effect sizes of the top 1000 and top 500 DEGs
}
