% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype_correlation.r
\name{plot_celltype_correlation}
\alias{plot_celltype_correlation}
\title{Computes the correlation of the log-foldchange of the DEGs (at specified p-value) for a given celltype, across all datasets}
\usage{
plot_celltype_correlation(dataset_name, allstudies, celltype, p_val = 1)
}
\arguments{
\item{dataset_name}{name of the dataset used to select significant DEGs from (specified as a string, name as in allstudies)}

\item{allstudies}{a list containing all the datasets (most likely as SCE objects)}

\item{celltype}{the celltype to compute correlation for}

\item{p_val}{the cut-off p-value which will be used to select DEGs (default is 1 to include all the genes)}
}
\value{
correlation matrix, plot and the number of DEGs at the specified p-value
}
\description{
Computes the correlation of the log-foldchange of the DEGs (at specified p-value) for a given celltype, across all datasets
}
