% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_de_analysis.R
\name{plot_de_analysis}
\alias{plot_de_analysis}
\title{Create differential expression analysis plots. Run by sc_cell_type_de()}
\usage{
plot_de_analysis(
  pb_dat,
  y,
  celltype_DEGs_dt,
  celltype_all_genes_dt,
  counts_celltypes,
  folder
)
}
\arguments{
\item{pb_dat}{A list containing
sumDat: matrix of the summed pseudobulk count values
annot_pb: dataframe of the annotation data from the SCE rolled up based on the pseudobulk aggregation.}

\item{y}{the column name in the SCE object for the return variable e.g. "diagnosis" - Case or disease. y can be discrete (logisitc regression) or continuous (linear regression)}

\item{celltype_DEGs_dt}{data table containing the DEGs for each cell type with their differential expression data}

\item{celltype_all_genes_dt}{data table containing the all genes for each cell type with their differential expression data}

\item{folder}{the folder where the graphs from the differential expression analysis are saved. Default will create a folder in the current working directory "sc.cell.type.de.graphs". False will skip plotting.
Creates and saves plot in relevant directory}

\item{celltype_counts}{vector with the counts of cells after QC in each cell type}
}
\description{
Create differential expression analysis plots. Run by sc_cell_type_de()
}
