% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_permutations.r
\name{random_permutations}
\alias{random_permutations}
\title{Obtain highly randomised permutations of a specified dataset (based on sex labels), while maintaining consistency with the sample ID}
\usage{
random_permutations(
  SCE,
  sampleID = "Donor.ID",
  sexID = "sex",
  Nrandom_perms = 5,
  Ntests = 100
)
}
\arguments{
\item{SCE}{the input data (should be an SCE object)}

\item{sampleID}{sample ID}

\item{sexID}{sex ID}

\item{Nrandom_perms}{the number of randomised permutations the user needs (default 5)}

\item{Ntests}{the number of random permutation tests (default 100)}
}
\value{
a list of size Nrandom_perms, with each item being a randomised dataset
}
\description{
Obtain highly randomised permutations of a specified dataset (based on sex labels), while maintaining consistency with the sample ID
}
