% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cells.r
\name{sample_cells}
\alias{sample_cells}
\title{Sample the data for a specified mean number of cells (per sample) from the dataset}
\usage{
sample_cells(SCE, meanCells, sampleID = "Donor.ID", Nperms = 20)
}
\arguments{
\item{SCE}{the input data (should be an SCE object)}

\item{meanCells}{the mean number of cells to be sampled from each individual}

\item{sampleID}{sample ID}

\item{Nperms}{the number of subsets the user needs (default 20)}
}
\value{
a list of size Nperms, with each item being a subset of the input data containing the specified mean number of cells per sample
}
\description{
Sample the data for a specified mean number of cells (per sample) from the dataset
}
