% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_individuals.r
\name{sample_individuals}
\alias{sample_individuals}
\title{Sample the data for a specified number of individuals from the dataset}
\usage{
sample_individuals(
  data,
  Nsamples,
  sampleID = "Donor.ID",
  sexID = "sex",
  Nperms = 20
)
}
\arguments{
\item{data}{the input data (should be an SCE object)}

\item{Nsamples}{the number of samples to be included in each subset}

\item{sampleID}{the column name in the SCE object which is the patient identifier (usually just "sample_id")}

\item{sexID}{the column name in the SCE object which is the sex (usually just "sex" or "Sex")}

\item{Nperms}{the number of subsets the user needs (default 20)}
}
\value{
a list of size Nperms, with each item being a subset of the input SCE containing the specified number of individuals
}
\description{
Sample the data for a specified number of individuals from the dataset
}
