% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_pairs.r
\name{subset_pairs}
\alias{subset_pairs}
\title{Obtain independent pairs of subsets of a specified dataset, based on sample ID}
\usage{
subset_pairs(data, sampleID = "Donor.ID", Noutputs = 5)
}
\arguments{
\item{data}{the input data (should be an SCE object)}

\item{sampleID}{sample ID}

\item{Noutputs}{the number of randomised permutations the user needs (default 5)}
}
\value{
a list of size Noutputs, with each item being a pair of independent datasets
}
\description{
Obtain independent pairs of subsets of a specified dataset, based on sample ID
}
