% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input_parameters_power.r
\name{validate_input_parameters_power}
\alias{validate_input_parameters_power}
\title{Tests input parameters for functions}
\usage{
validate_input_parameters_power(
  SCE = "placeholder",
  range_downsampled = "placeholder",
  output_path = "placeholder",
  inpath = "placeholder",
  sampled = "placeholder",
  sampleID = "placeholder",
  design = "placeholder",
  sexID = "placeholder",
  celltypeID = "placeholder",
  assay_name = "placeholder",
  coef = "placeholder",
  fdr = "placeholder",
  nom_pval = "placeholder",
  Nperms = "placeholder",
  N_randperms = "placeholder",
  N_subsetpairs = "placeholder",
  y = "placeholder",
  region = "placeholder",
  control = "placeholder",
  pval_adjust_method = "placeholder",
  rmv_zero_count_genes = "placeholder",
  abs_effect_size_thresholds = "placeholder",
  upreg_effect_size_thresholds = "placeholder",
  downreg_effect_size_thresholds = "placeholder"
)
}
\arguments{
\item{SCE}{the input data (should be an SCE object)}

\item{range_downsampled}{vector or list containing values which the SCE will be downsampled at, in ascending order}

\item{output_path}{base path in which outputs will be stored}

\item{inpath}{base path where downsampled DGE analysis output is stored (taken to be output_path if not provided)}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{sampleID}{sample ID}

\item{design}{the design formula of class type \code{formula}. Equation used to fit the model- data for the generalised linear model e.g. expression ~ sex + pmi + disease}

\item{sexID}{sex ID}

\item{celltypeID}{cell type ID}

\item{assay_name}{the name of the assay in the SCE object to be used for the analysis (default is "counts")}

\item{coef}{which coefficient to carry out DE analysis with respect to}

\item{fdr}{the cut-off False Discovery Rate below which to select DEGs}

\item{nom_pval}{the cut-off nominal P-value below which to select DEGs (as an alternative to FDR)}

\item{Nperms}{number of subsets created when downsampling at each level}

\item{N_randperms}{number of randomised permutations of the dataset (based on sex) to be correlated}

\item{N_subsetpairs}{number of pairs of subsets of the dataset to be correlated}

\item{y}{the column name in the SCE object for the return variable e.g. "diagnosis" - Case or disease. Default is the last variable in the design formula. y can be discrete (logistic regression) or continuous (linear regression)}

\item{region}{the column name in the SCE object for the study region. If there are multiple regions in the study (for example two brain regions). Pseudobulk values can be derived separately. Default is "single_region" which will not split by region.}

\item{control}{character specifying which control level for the differential expression analysis e.g. in a case/control/other study use "control" in the y column to compare against. NOTE only need to specify if more than two groups in y, leave as default value for two groups or continuous y. Default is NULL.}

\item{pval_adjust_method}{the adjustment method for the p-value in the differential expression analysis. Default is benjamini hochberg "BH". See  stats::p.adjust for available options}

\item{rmv_zero_count_genes}{whether genes with no count values in any cell should be removed. Default is TRUE}

\item{abs_effect_size_thresholds}{Optional. Numeric vector of effect size (absolute logFC) thresholds to use for power analysis. If not provided, defaults to 25th, 50th and 75h percentiles of the absolute logFCs. Must contain non-negative, increasing values.}

\item{upreg_effect_size_thresholds}{Optional. Numeric vector of effect size thresholds to use for power analysis (for up-regulated DEGs). If not provided, defaults to 25th, 50th and 75h percentiles of the positive logFCs. Must contain non-negative, increasing values.}

\item{downreg_effect_size_thresholds}{Optional. Numeric vector of effect size thresholds to use for power analysis (for down-regulated DEGs). If not provided, defaults to 25th, 50th and 75h percentiles of the negative logFCs. Must contain negative (or zero), increasing values.
Checks all power analysis parameters are specified correctly}
}
\description{
Tests input parameters for functions
}
