% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_downsampling_DGEanalysis.r
\name{bulk_downsampling_DGEanalysis}
\alias{bulk_downsampling_DGEanalysis}
\title{Downsample the dataset, based either on the individuals or cells, and run DE analysis on each downsampled output. Save results in a dataframe}
\usage{
bulk_downsampling_DGEanalysis(
  datasets,
  celltype_correspondence,
  sampled = "individuals",
  sampleID = "donor_id",
  celltype_ID = "cell_type",
  path = "placeholder"
)
}
\arguments{
\item{datasets}{list of the input data (elements should be SCE objects)}

\item{celltype_correspondence}{list of different names specifying each cell type}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{sampleID}{sample ID}

\item{celltype_ID}{cell type ID}

\item{path}{base output directory where down-sampled DGE analysis outputs will be saved
Saves DGE analysis output in the correct directory, to be used by other bulk analysis functions}
}
\description{
Downsample the dataset, based either on the individuals or cells, and run DE analysis on each downsampled output. Save results in a dataframe
}
