% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_downsampling_DGEanalysis.r
\name{bulk_downsampling_DGEanalysis}
\alias{bulk_downsampling_DGEanalysis}
\title{Downsample the dataset, based either on the individuals or cells, and run DE analysis on each downsampled output. Save results in a dataframe}
\usage{
bulk_downsampling_DGEanalysis(
  SCEs,
  dataset_names,
  celltype_correspondence,
  sampled = "individuals",
  sampleIDs = "donor_id",
  celltypeIDs = "cell_type",
  output_path = getwd(),
  pvalue = 0.05,
  Nperms = 20
)
}
\arguments{
\item{SCEs}{list of the input data (elements should be SCE objects)}

\item{dataset_names}{list of the names of the datasets (as you would like them to appear in the "output_path" directory)}

\item{celltype_correspondence}{list of different names specifying each cell type, where the order of celltype_names correspond to the order of SCEs}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{sampleIDs}{list or vector of sample IDs (in order of SCEs)}

\item{celltypeIDs}{list or vector of cell type IDs (in order of SCEs)}

\item{output_path}{base output directory where down-sampled DGE analysis outputs will be saved}

\item{pvalue}{the cut-off p-value used to select DEGs}

\item{Nperms}{number of permutations of DGE analysis outputs for each sample
Saves DGE analysis output in the correct directory, to be used by other bulk analysis functions}
}
\description{
Downsample the dataset, based either on the individuals or cells, and run DE analysis on each downsampled output. Save results in a dataframe
}
