% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_downsampling_range.r
\name{bulk_downsampling_range}
\alias{bulk_downsampling_range}
\title{Obtain the range of values to downsample at (either for individuals, or mean number of cells per individual), for bulk analysis}
\usage{
bulk_downsampling_range(
  SCEs,
  celltype_correspondence,
  sampled = "individuals",
  sampleID = "donor_id",
  celltypeID = "cell_type"
)
}
\arguments{
\item{SCEs}{list of the input data (elements should be SCE objects)}

\item{celltype_correspondence}{list of different names specifying each cell type}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{sampleID}{sample ID}

\item{celltypeID}{cell type ID}
}
\value{
list containing values which the data will be downsampled at, in ascending order
}
\description{
Obtain the range of values to downsample at (either for individuals, or mean number of cells per individual), for bulk analysis
}
