% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_power_analysis.r
\name{bulk_power_analysis}
\alias{bulk_power_analysis}
\title{Runs entire bulk RNA-seq power analysis pipeline}
\usage{
bulk_power_analysis(
  SCEs,
  dataset_names,
  celltype_correspondence,
  output_path = getwd(),
  celltypeIDs = "cell_type",
  sampled = "individuals",
  sampleIDs = "donor_id",
  bulkDE = "placeholder",
  bulk_cutoff = 0.9,
  pvalue = 0.05,
  Nperms = 20,
  fontsize_axislabels = 12,
  fontsize_axisticks = 9,
  fontsize_title = 14,
  fontsize_legendlabels = 9,
  fontsize_legendtitle = 9,
  plot_title = "placeholder"
)
}
\arguments{
\item{SCEs}{list of the input data (elements should be SCE objects)}

\item{dataset_names}{list of the names of the datasets (as you would like them to appear in the "output_path" directory)}

\item{celltype_correspondence}{list of different names specifying each cell type}

\item{output_path}{path storing the down-sampled DGE analysis outputs for each dataset (and saves outputs)}

\item{celltypeIDs}{list or vector of cell type IDs (in order of SCEs)}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{sampleIDs}{list or vector of sample IDs (in order of SCEs)}

\item{bulkDE}{DGE analysis output for a bulk RNA-seq dataset: rows (rownames) should be the genes, columns should be tissues, and entries should be significance levels}

\item{bulk_cutoff}{percentage (proportion), specified so that we select DEGs common across >=bulk_cutoff of the tissues in the Bulk dataset}

\item{pvalue}{the cut-off p-value used to select DEGs (for bulk data)}

\item{Nperms}{number of permutations of DGE analysis outputs for each sample}

\item{fontsize_axislabels}{font size for axis labels in plot}

\item{fontsize_axisticks}{font size for axis tick labels in plot}

\item{fontsize_title}{font size for plot title}

\item{fontsize_legendlabels}{font size for legend labels in plot}

\item{fontsize_legendtitle}{font size for legend title in plot}

\item{plot_title}{plot title
Saves all plots in the appropriate directories}
}
\description{
Runs entire bulk RNA-seq power analysis pipeline
}
\examples{
\dontrun{
# Example of celltype_correspondence:
celltype_correspondence <- list([overall_celltype_name1] = c([celltype1.1], [celltype1.2], ...),
                                [overall_celltype_name2] = c([celltype2.1], [celltype2.2], ...),
                                ...)
# E.g.
celltype_correspondence <- list("Microglia" = c("micro", "microglial cells", "Micro"),
                               "Astrocytes" = c("Astrocytes", "astro"),
                               "Oligodendrocytes" = c("oligo", "Oligodendrocytes"))
# If a celltype is not present in the dataset, the corresponding value should be set to NA. E.g.:
celltype_correspondence <- list("Microglia" = c("micro",NA), "Endothelial" = c(NA,"endo"))
}
}
