% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_power_analysis.r
\name{bulk_power_analysis}
\alias{bulk_power_analysis}
\title{Runs entire bulk RNA-seq power analysis pipeline}
\usage{
bulk_power_analysis(
  datasets,
  celltype,
  celltype_correspondence,
  path = getwd(),
  range_downsampled = "placeholder",
  celltype_ID = "cell_type",
  sampled = "individuals",
  sampleID = "donor_id",
  bulkDE = "placeholder",
  bulk_cutoff = 0.9,
  pvalue = 0.05,
  Nperms = 20,
  fontsize_axislabels = 12,
  fontsize_axisticks = 9,
  fontsize_title = 14,
  fontsize_legendlabels = 9,
  fontsize_legendtitle = 9,
  plot_title = "placeholder"
)
}
\arguments{
\item{datasets}{list of the input data (elements should be SCE objects)}

\item{celltype}{the cell type we are focusing on (name as it appears in cell type sub-directory name)}

\item{celltype_correspondence}{list of different names specifying each cell type}

\item{path}{path storing the down-sampled DGE analysis outputs for each dataset (and saves outputs)}

\item{range_downsampled}{vector or list containing values which the data will be downsampled at, in ascending order}

\item{celltype_ID}{cell type ID}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{sampleID}{sample ID}

\item{bulkDE}{DGE analysis output for a bulk RNA-seq dataset: rows (rownames) should be the genes, columns should be tissues, and entries should be significance levels}

\item{bulk_cutoff}{percentage (proportion), specified so that we select DEGs common across >=bulk_cutoff of the tissues in the Bulk dataset}

\item{pvalue}{the cut-off p-value used to select DEGs (for bulk data)}

\item{Nperms}{number of permutations of DGE analysis outputs for each sample}

\item{fontsize_axislabels}{font size for axis labels in plot}

\item{fontsize_axisticks}{font size for axis tick labels in plot}

\item{fontsize_title}{font size for plot title}

\item{fontsize_legendlabels}{font size for legend labels in plot}

\item{fontsize_legendtitle}{font size for legend title in plot}

\item{plot_title}{plot title
Saves all plots in the appropriate directories}
}
\description{
Runs entire bulk RNA-seq power analysis pipeline
}
