% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.r
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Runs correlation analysis pipeline}
\usage{
correlation_analysis(
  main_dataset,
  SCEs,
  sampleIDs,
  celltypeIDs,
  celltype_correspondence,
  dataset_names,
  pvals = c(0.05, 0.025, 0.01, 0.001, 1e-04),
  alphaval = 0.25,
  N_randperms = 5,
  N_subsets = 5,
  sex_DEGs = FALSE,
  fontsize_yaxislabels = 12,
  fontsize_yaxisticks = 9,
  fontsize_title = 14,
  fontsize_legendlabels = 9,
  fontsize_legendtitle = 9,
  fontsize_facet_labels = 9,
  output_path = getwd()
)
}
\arguments{
\item{main_dataset}{name of the dataset used to select significant DEGs from (specified as a string, name as in dataset_names)}

\item{SCEs}{list of the input data (elements should be SCE objects)}

\item{sampleIDs}{list or vector of sample IDs (in order of SCEs)}

\item{celltypeIDs}{list or vector of cell type IDs (in order of SCEs)}

\item{celltype_correspondence}{list of different names specifying each cell type}

\item{dataset_names}{names of the datasets as they appear in the correlation plot (in order of SCEs)}

\item{pvals}{list of p-value cut-offs which will be used to select DEGs}

\item{alphaval}{(alpha) transparency of the non-mean boxplots}

\item{N_randperms}{number of random permutations of the dataset used to select significant DEGs from}

\item{N_subsets}{number of pairs of random subsets of the dataset used to select significant DEGs from}

\item{sex_DEGs}{If TRUE, only keep genes present on sex chromosmomes. Queries hspanies gene Ensembl dataset.}

\item{fontsize_yaxislabels}{font size for axis labels in plot}

\item{fontsize_yaxisticks}{font size for axis tick labels in plot}

\item{fontsize_title}{font size for plot title}

\item{fontsize_legendlabels}{font size for legend labels in plot}

\item{fontsize_legendtitle}{font size for legend title in plot}

\item{fontsize_facet_labels}{font size for facet labels}

\item{output_path}{base path in which outputs will be stored
Saves all plots and DGE analysis outputs in the appropriate directories}
}
\description{
Runs correlation analysis pipeline
}
\examples{
\dontrun{
# Example of celltype_correspondence:
celltype_correspondence <- list([overall_celltype_name1] = c([celltype1.1], [celltype1.2], ...),
                                [overall_celltype_name2] = c([celltype2.1], [celltype2.2], ...),
                                ...)
# E.g.
celltype_correspondence <- list("Microglia" = c("micro", "microglial cells", "Micro"),
                               "Astrocytes" = c("Astrocytes", "astro", NA),
                               "Oligodendrocytes" = c("oligo", "Oligodendrocytes", NA))

# Runnable example with bundled data
# 1. Prepare DGE
micro_tsai <- system.file("extdata", "Tsai_Micro.qs", package="poweranalysis")
SCE_tsai <- qs::qread(micro_tsai)
DGE_tsai <- poweranalysis::DGE_analysis(
    SCE_tsai,
    design = ~ sex,
    celltypeID="cluster_celltype",
    sampleID = "sample_id",
    coef = "M"
)
celltype_names <- list("Microglia" = list("Micro"))

# 2. Run correlation analysis
correlation_analysis(
    main_dataset = "tsai",
    SCEs = list(tsai),
    celltype_correspondence = celltype_names,
    dataset_names = list("tsai"),
    sex_DEGs = TRUE,    # Keep only genes on sex chromosomes
    output_path = output_path
)
}
}
