% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_boxplots.r
\name{correlation_boxplots}
\alias{correlation_boxplots}
\title{Obtain box plots for the correlations of all celltypes, and the mean correlations at a specified cutoff p-value}
\usage{
correlation_boxplots(
  corrMats,
  numRealDatasets,
  pvals,
  alphaval = 0.25,
  numPerms = 5,
  numSubsets = 5,
  sexDEGs = FALSE,
  fontsize_yaxislabels = 12,
  fontsize_yaxisticks = 9,
  fontsize_title = 14,
  fontsize_legendlabels = 9,
  fontsize_legendtitle = 9,
  fontsize_facet_labels = 9,
  output_path = getwd()
)
}
\arguments{
\item{corrMats}{(named) list of correlation matrices for each celltype with the final element being the mean correlation matrix, all at specified p-value}

\item{numRealDatasets}{total number of \emph{real} datasets (most likely the number of studies, but sometimes a study may be split e.g. into 2 brain regions, so in this case it would be the number of studies plus 1)}

\item{pvals}{the cut-off p-value which was used to select DEGs}

\item{alphaval}{(alpha) transparency of the non-mean boxplots}

\item{numPerms}{number of random permutations of the dataset used to select significant DEGs from}

\item{numSubsets}{number of pairs of random subsets of the dataset used to select significant DEGs from}

\item{sexDEGs}{true if DEGs come from sex chromosomes, else false}

\item{fontsize_yaxislabels}{font size for axis labels in plot}

\item{fontsize_yaxisticks}{font size for axis tick labels in plot}

\item{fontsize_title}{font size for plot title}

\item{fontsize_legendlabels}{font size for legend labels in plot}

\item{fontsize_legendtitle}{font size for legend title in plot}

\item{fontsize_facet_labels}{font size for facet labels}

\item{output_path}{base path in which outputs will be stored}
}
\value{
box plots for correlation matrices at a certain p-value cut-off, sorted by celltype and then type of correlation
}
\description{
Obtain box plots for the correlations of all celltypes, and the mean correlations at a specified cutoff p-value
}
