% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling_range.r
\name{downsampling_range}
\alias{downsampling_range}
\title{Obtain the range of values to downsample at (either for individuals, or mean number of cells per individual)}
\usage{
downsampling_range(data, sampled = "individuals", sampleID = "Donor.ID")
}
\arguments{
\item{data}{the input data (should be an SCE object)}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{sampleID}{sample ID}
}
\value{
list containing values which the data will be downsampled at, in ascending order
}
\description{
Obtain the range of values to downsample at (either for individuals, or mean number of cells per individual)
}
