% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_celltype_DEGs.r
\name{gather_celltype_DEGs}
\alias{gather_celltype_DEGs}
\title{Collate DEGs detected in DGE analysis outputs, across all celltypes in a dataset (datasets/DGE analysis outputs should have common celltypes as specified below)}
\usage{
gather_celltype_DEGs(
  celltype_correspondence,
  pvalue = 0.05,
  Nperms = 20,
  output_path = getwd()
)
}
\arguments{
\item{celltype_correspondence}{A named vector that maps a standard cell type label (e.g., \code{"Endo"}, \code{"Micro"}) to how that cell type appears in each dataset. Use \code{NA} if the cell type is not present in a given dataset.}

\item{pvalue}{Numeric. P-value threshold for defining DEGs in the bulk dataset.}

\item{Nperms}{Number of permutations to perform for each down-sampling level. Default is 20; for quick testing, use a smaller value (e.g., \code{Nperms = 3}).}

\item{output_path}{A directory path where down-sampled outputs and plots will be saved.
Saves combined list of DEGs (across all cell types) in a subdirectory inside the dataset directory}
}
\description{
Collate DEGs detected in DGE analysis outputs, across all celltypes in a dataset (datasets/DGE analysis outputs should have common celltypes as specified below)
}
