% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_celltype_DEGs.r
\name{gather_celltype_DEGs}
\alias{gather_celltype_DEGs}
\title{Collate DEGs detected in DGE analysis outputs, across all celltypes in a dataset (datasets/DGE analysis outputs should have common celltypes as specified below)}
\usage{
gather_celltype_DEGs(
  celltype_correspondence,
  pvalue = 0.05,
  Nperms = 20,
  output_path = getwd()
)
}
\arguments{
\item{celltype_correspondence}{list of different names specifying each cell type}

\item{pvalue}{the cut-off p-value used to select DEGs}

\item{Nperms}{number of permutations of DGE analysis outputs for each sample}

\item{output_path}{path storing the down-sampled DGE analysis for each single-cell dataset, generated for bulk analysis
Saves combined list of DEGs (across all cell types) in a subdirectory inside the dataset directory}
}
\description{
Collate DEGs detected in DGE analysis outputs, across all celltypes in a dataset (datasets/DGE analysis outputs should have common celltypes as specified below)
}
