% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pseudobulk.R
\name{make_pseudobulk}
\alias{make_pseudobulk}
\title{Calculate the summed pseudobulk values for an SCE object based on one single cell type only. Ensure to filter SCE to pass one cell type's data.}
\usage{
make_pseudobulk(
  data,
  pseudobulk_ID,
  pb_columns = NULL,
  region = "single_region",
  rmv_zero_count_genes = TRUE
)
}
\arguments{
\item{data}{SingleCellExperiment object, a specialised S4 class for storing data from single-cell experiments.}

\item{pseudobulk_ID}{Column name in the SCE object to perform pseudobulk on, usually the patient identifier. This column is used for grouping in the pseudobulk approach}

\item{pb_columns}{Vector, list of annotation column names in the SCE object to be returned in annot_pb rolled up to pseudobulk level. Default is NULL which won't return any information in annot_pb.}

\item{region}{Column name in the SCE object for the study region. If there are multiple regions in the study (for example two brain regions). Pseudobulk values can be derived separately. Default is "single_region" which will not split by region.}

\item{rmv_zero_count_genes}{Whether genes with no count values in any cell should be removed. Default is TRUE}
}
\value{
a list containing:
sumDat: matrix of the summed pseudobulk count values
annot_pb: dataframe of the annotation data from the SCE rolled up based on the pseudobulk aggregation.
}
\description{
Calculate the summed pseudobulk values for an SCE object based on one single cell type only. Ensure to filter SCE to pass one cell type's data.
}
