% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype_correlation.r
\name{plot_celltype_correlation}
\alias{plot_celltype_correlation}
\title{Computes the correlation of the log-foldchange of the DEGs (at specified p-value) for a given celltype, across all datasets}
\usage{
plot_celltype_correlation(
  main_dataset,
  DEouts,
  celltypes_list,
  dataset_names,
  sex_DEGs = FALSE,
  pval = 1
)
}
\arguments{
\item{main_dataset}{name of the dataset used to select significant DEGs from (specified as a string, name as in dataset_names)}

\item{DEouts}{a list containing outputs of DGE analysis (as returned/optionally saved by DGE_analysis) for datasets to be used in the correlation analysis}

\item{celltypes_list}{list of different names specifying each cell type (in order for each dataset in DEouts)}

\item{dataset_names}{names of the datasets as they will appear in the correlation plot}

\item{sex_DEGs}{If TRUE, only keep genes present on sex chromosmomes. Queries hspanies gene Ensembl dataset.}

\item{pval}{the cut-off p-value which will be used to select DEGs (default is 1 to include all the genes)}
}
\value{
correlation matrix, plot and the number of DEGs at the specified p-value
}
\description{
Computes the correlation of the log-foldchange of the DEGs (at specified p-value) for a given celltype, across all datasets
}
