% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_correlation.r
\name{plot_mean_correlation}
\alias{plot_mean_correlation}
\title{Obtain the average correlation (across celltypes) at a specified cutoff p-value}
\usage{
plot_mean_correlation(
  dataset_name,
  DEouts,
  pvals,
  celltype_correspondence,
  data_names,
  output_path = getwd()
)
}
\arguments{
\item{dataset_name}{name of the dataset used to select significant DEGs from (specified as a string, name as in DEouts)}

\item{DEouts}{a list containing outputs of DGE analysis (as returned/optionally saved by DGE_analysis) for datasets to be used in the correlation analysis}

\item{pvals}{the cut-off p-value which will be used to select DEGs}

\item{celltype_correspondence}{list of different names specifying each cell type}

\item{data_names}{names of the datasets as they appear in the correlation plot}

\item{output_path}{base path in which outputs will be stored
Saves mean correlation matrix (and actual values) in the appropriate directory}
}
\description{
Obtain the average correlation (across celltypes) at a specified cutoff p-value
}
