% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_correlation.r
\name{plot_mean_correlation}
\alias{plot_mean_correlation}
\title{Obtain the average correlation (across celltypes) at a specified cutoff p-value}
\usage{
plot_mean_correlation(
  main_dataset,
  SCEs,
  dataset_names,
  celltype_correspondence,
  sampleIDs = "donor_id",
  celltypeIDs = "cell_type",
  assay_names = "counts",
  N_randperms = 5,
  N_subsets = 5,
  pvals = c(0.05, 0.025, 0.01, 0.001, 1e-04),
  sex_DEGs = FALSE,
  output_path = getwd()
)
}
\arguments{
\item{main_dataset}{name of the dataset used to select significant DEGs from, and for which random permutations and subset pairs will be created (to be specified as a string, name as in dataset_names)}

\item{SCEs}{list of the input data (elements should be SCE objects)}

\item{dataset_names}{names of the datasets as they appear in the correlation plot (in order of SCEs)}

\item{celltype_correspondence}{list of different names specifying each cell type}

\item{sampleIDs}{list or vector of sample IDs (in order of SCEs)}

\item{celltypeIDs}{list or vector of cell type IDs (in order of SCEs)}

\item{assay_names}{names of the assays in each SCE that will be used for the analysis (in order of SCEs). Default is a vector with all entries "counts", which uses the count assay in each SCE.}

\item{N_randperms}{number of random permutations of the dataset to be created}

\item{N_subsets}{number of pairs of random subsets of the dataset to be created}

\item{pvals}{the cut-off p-value which will be used to select DEGs}

\item{sex_DEGs}{If TRUE, only keep genes present on sex chromosmomes. Queries hspanies gene Ensembl dataset.}

\item{output_path}{base path in which outputs will be stored
Saves mean correlation matrix (and actual values) in the appropriate directory}
}
\description{
Obtain the average correlation (across celltypes) at a specified cutoff p-value
}
