% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_bulk_DEGs_sc.r
\name{prop_bulk_DEGs_sc}
\alias{prop_bulk_DEGs_sc}
\title{Obtain overall percentage overlap between DEGs from bulk data (DE across all tissues) and various scRNA-seq datasets, across all common cell types}
\usage{
prop_bulk_DEGs_sc(
  bulkDE,
  bulk_cutoff = 0.9,
  pvalue = 0.05,
  sampled = "individuals",
  fontsize_axislabels = 12,
  fontsize_axisticks = 9,
  fontsize_title = 14,
  fontsize_legendlabels = 9,
  fontsize_legendtitle = 9,
  plot_title = "placeholder",
  output_path = getwd()
)
}
\arguments{
\item{bulkDE}{DGE analysis output for a bulk RNA-seq dataset (e.g., \code{LFSR.tsv}): rows (rownames) should be the genes, columns should be tissues, and entries should be significance levels}

\item{bulk_cutoff}{Numeric. Proportion (0–1) of bulk tissues in which a gene must be differentially expressed to be considered (e.g., 0.9 selects DEGs found in ≥90\% of tissues).}

\item{pvalue}{Numeric. P-value threshold for defining DEGs in the bulk dataset.}

\item{sampled}{Specifies the unit of down-sampling. Can be either \code{"individuals"} or \code{"cells"}, depending on whether the analysis downsamples across samples or cells.}

\item{fontsize_axislabels}{font size for axis labels in plot}

\item{fontsize_axisticks}{font size for axis tick labels in plot}

\item{fontsize_title}{font size for plot title}

\item{fontsize_legendlabels}{font size for legend labels in plot}

\item{fontsize_legendtitle}{font size for legend title in plot}

\item{plot_title}{plot title}

\item{output_path}{A directory path where down-sampled outputs and diagnostic plots will be saved.
Saves plot showing percentage DEGs from bulk data found in each scRNA-seq dataset, in the appropriate directory}
}
\description{
Obtain overall percentage overlap between DEGs from bulk data (DE across all tissues) and various scRNA-seq datasets, across all common cell types
}
