% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input_parameters_bulk.r
\name{validate_input_parameters_bulk}
\alias{validate_input_parameters_bulk}
\title{Tests input parameters for functions}
\usage{
validate_input_parameters_bulk(
  bulkDE = "placeholder",
  inpath = "placeholder",
  range_downsampled = "placeholder",
  celltype = "placeholder",
  sampled = "placeholder",
  bulk_cutoff = "placeholder",
  pvalue = "placeholder",
  celltype_names = "placeholder",
  datasets2 = "placeholder",
  celltype_names2 = "placeholder",
  Nperms = "placeholder",
  fontsize_axislabels = "placeholder",
  fontsize_axisticks = "placeholder",
  fontsize_title = "placeholder",
  fontsize_legendlabels = "placeholder",
  fontsize_legendtitle = "placeholder",
  plot_title = "placeholder"
)
}
\arguments{
\item{bulkDE}{DGE analysis output for a bulk RNA-seq dataset: rows (rownames) should be the genes, columns should be tissues, and entries should be significance levels}

\item{inpath}{path storing the down-sampled DGE analysis for each single-cell dataset, generated for bulk analysis}

\item{range_downsampled}{vector or list containing values which the data will be downsampled at, in ascending order}

\item{celltype}{the cell type we are focusing on (name as it appears in cell type sub-directory name)}

\item{sampled}{downsampling carried out based on what (either "individuals" or "cells")}

\item{bulk_cutoff}{percentage (proportion between 0 and 1), specified so that we select DEGs common across >= bulk_cutoff of the tissues in the Bulk dataset}

\item{pvalue}{the cut-off p-value used to select DEGs (for both, bulk and scRNA-seq datasets)}

\item{celltype_names}{list of the names specifying cell types in each DGE analysis output (in order they appear in the directory)}

\item{datasets2}{list of datasets to be used, with celltype names (in DGE analysis outputs) as celltype_names2}

\item{celltype_names2}{alt. list of the names specifying cell types in each DGE analysis output (in order they appear in the directory), if all are not the same as celltype}

\item{Nperms}{number of permutations of DGE analysis outputs for each sample}

\item{fontsize_axislabels}{font size for axis labels in plot}

\item{fontsize_axisticks}{font size for axis tick labels in plot}

\item{fontsize_title}{font size for plot title}

\item{fontsize_legendlabels}{font size for legend labels in plot}

\item{fontsize_legendtitle}{font size for legend title in plot}

\item{plot_title}{plot title
Checks all bulk analysis parameters are specified correctly}
}
\description{
Tests input parameters for functions
}
